refactor(chat): migrate chat components to new chat-ui directory
- Updated imports in page and dashboard shell to reference new chat-ui components.
- Removed deprecated chat-v2 components including ChatShell, Composer, DocumentViewerSheet, EmptyState, FacultyDashboard, Header, Message, MessageList, and ProfileModal.
- This refactor aims to streamline the chat feature organization and improve maintainability.
