+++ aura/components/features/chat-ui/ChatShell.tsx
