+++ aura/components/features/chat-ui/EmptyState.tsx
