+++ aura/components/features/chat-ui/FacultyDashboard.tsx
