+++ aura/components/features/chat-ui/Header.tsx
