+++ aura/components/features/chat-ui/ProfileModal.tsx
