+++ aura/components/features/chat-ui/Sidebar.tsx
