+++ aura/components/features/chat-ui/StreamingIndicator.tsx
