+++ aura/components/features/chat-ui/StudentDashboard.tsx
